+++ pkg/server/web/web.go
@@ -15,6 +15,11 @@
 package web
 
 import (
+	"compress/gzip"
+	"crypto/md5"
+	"encoding/hex"
+	"fmt"
+	"io"
 	"io/fs"
 	"net/http"
 	"strings"
@@ -26,6 +31,7 @@ import (
 type Server struct {
 	spaFS        http.FileSystem
 	indexContent []byte
+	indexETag    string
 }
 
 func NewServer() (*Server, error) {
@@ -51,9 +57,14 @@ func NewServer() (*Server, error) {
 		return nil, err
 	}
 
+	// Generate ETag for index.html
+	hash := md5.Sum(indexContent)
+	indexETag := hex.EncodeToString(hash[:])
+
 	return &Server{
 		spaFS:        http.FS(subFS),
 		indexContent: indexContent,
+		indexETag:    indexETag,
 	}, nil
 }
 
@@ -64,33 +75,82 @@ func (s *Server) ServeSPA(w http.ResponseWriter, r *http.Request) {
 	if err == nil {
 		defer f.Close()
 
-		if isStaticFile(path) {
-			w.Header().Set("Cache-Control", "public, max-age=31536000")
-			w.Header().Set("Expires", time.Now().Add(365*24*time.Hour).Format(time.RFC1123))
+		info, err := f.Stat()
+		if err != nil {
+			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+			return
 		}
 
+		etag := fmt.Sprintf(`"%x-%x"`, info.Size(), info.ModTime().Unix())
+
+		lastModified := info.ModTime().Format(http.TimeFormat)
+
+		w.Header().Set("ETag", etag)
+		w.Header().Set("Last-Modified", lastModified)
+
+		if matchETag := r.Header.Get("If-None-Match"); matchETag != "" {
+			if matchETag == etag {
+				w.WriteHeader(http.StatusNotModified)
+				return
+			}
+		}
+
+		if modifiedSince := r.Header.Get("If-Modified-Since"); modifiedSince != "" {
+			if t, err := time.Parse(http.TimeFormat, modifiedSince); err == nil {
+				if info.ModTime().Unix() <= t.Unix() {
+					w.WriteHeader(http.StatusNotModified)
+					return
+				}
+			}
+		}
+
+		if strings.HasSuffix(path, ".js") || strings.HasSuffix(path, ".css") ||
+			strings.HasSuffix(path, ".png") || strings.HasSuffix(path, ".jpg") ||
+			strings.HasSuffix(path, ".svg") || strings.HasSuffix(path, ".woff2") {
+			w.Header().Set("Cache-Control", "public, max-age=31536000, immutable") // 1 year
+		} else {
+			w.Header().Set("Cache-Control", "public, max-age=3600") // 1 hour
+		}
+
+		w.Header().Set("Transfer-Encoding", "chunked")
+
+		// Serve the file
 		http.FileServer(s.spaFS).ServeHTTP(w, r)
 		return
 	}
+	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+	w.Header().Set("ETag", `"`+s.indexETag+`"`)
 
-	w.Header().Set("Cache-Control", "no-cache, no-store, must-revalidate")
+	if r.Header.Get("If-None-Match") == `"`+s.indexETag+`"` {
+		w.WriteHeader(http.StatusNotModified)
+		return
+	}
+
+	w.Header().Set("Cache-Control", "no-cache")
 	w.Header().Set("Pragma", "no-cache")
 	w.Header().Set("Expires", "0")
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+
 	w.WriteHeader(http.StatusOK)
 	w.Write(s.indexContent)
 }
 
-func isStaticFile(path string) bool {
-	extensions := []string{".js", ".css", ".png", ".jpg", ".jpeg", ".gif", ".svg", ".ico", ".woff", ".woff2", ".ttf", ".eot"}
-	for _, ext := range extensions {
-		if strings.HasSuffix(path, ext) {
-			return true
-		}
-	}
-	return false
+type gzipResponseWriter struct {
+	io.Writer
+	http.ResponseWriter
+}
+
+func (w gzipResponseWriter) Write(b []byte) (int, error) {
+	return w.Writer.Write(b)
 }
 
 func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	if strings.Contains(r.Header.Get("Accept-Encoding"), "gzip") {
+		w.Header().Set("Content-Encoding", "gzip")
+		gz := gzip.NewWriter(w)
+		defer gz.Close()
+		s.ServeSPA(gzipResponseWriter{Writer: gz, ResponseWriter: w}, r)
+		return
+	}
+
 	s.ServeSPA(w, r)
 }