+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+Fix typo `mesure` instead of `measure`
+
+## [0.18.0] - 2025-04-30
+
+### Added
+
+- Static files are served using GZip
+- Static fiels are served with ETag and Cache header fields
+
+### Fixed
+
+- Entrypoint JS/CSS has no chunk hash
+
+## [0.17.0] = 2025-04-29
+
+### Added
+
+- Add policy unlogged sign
+
 ## [0.16.0] = 2025-04-29
 
 ### Added